pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatchling",
]

[project]
name = "django-imagefield"
description = "Image field for Django"
readme = "README.rst"
license = { text = "BSD-3-Clause" }
authors = [
  { name = "Matthias Kestenholz", email = "mk@feinheit.ch" },
]
requires-python = ">=3.10"
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Software Development",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dynamic = [
  "version",
]
dependencies = [
  "pillow>=9,!=11",
]
optional-dependencies.tests = [
  "coverage",
  "pillow",
]
optional-dependencies.vips = [
  "pyvips[binary]>=2.2",
]
urls.Homepage = "https://github.com/matthiask/django-imagefield/"

[tool.hatch.build]
include = [
  "imagefield/",
]

[tool.hatch.version]
path = "imagefield/__init__.py"

[tool.ruff]
target-version = "py310"

fix = true
show-fixes = true
lint.extend-select = [
  # flake8-bugbear
  "B",
  # flake8-comprehensions
  "C4",
  # mmcabe
  "C90",
  # flake8-django
  "DJ",
  "E",
  # pyflakes, pycodestyle
  "F",
  # flake8-boolean-trap
  "FBT",
  # isort
  "I",
  # flake8-gettext
  "INT",
  # flake8-implicit-string-concatenation
  "ISC",
  # pygrep-hooks
  "PGH",
  # flake8-pie
  "PIE",
  # pylint
  "PLC",
  "PLE",
  "PLW",
  # unused noqa
  "RUF100",
  # pyupgrade
  "UP",
  "W",
  # flake8-2020
  "YTT",
]
lint.extend-ignore = [
  # Allow zip() without strict=
  "B905",
  # No line length errors
  "E501",
  # Percent-formatting is fine
  "UP031",
]
lint.per-file-ignores."*/migrat*/*" = [
  # Allow using PascalCase model names in migrations
  "N806",
  # Ignore the fact that migration files are invalid module names
  "N999",
]
lint.per-file-ignores."tests/*" = [
  # Who cares about __str__ in the test suite
  "DJ008",
]
lint.isort.combine-as-imports = true
lint.isort.lines-after-imports = 2
lint.mccabe.max-complexity = 15
